refactor: Deleted unnecessary files and improve naming
