+++ README.md
@@ -28,6 +28,15 @@ To use AURA, follow these steps:
 2. Install the required dependencies:
 
     ```bash
+    # Setup ollama for AI integration
+    curl -fsSL https://ollama.com/install.sh | sh
+    
+    # Get the model
+    ollama pull llama3
+    
+    ```
+
+    ```bash
     pip install -r requirements.txt
     ```
